グループ圧　表示位置調整　他

+++ FLEX.NET/clsPLC.vb
@@ -1067,7 +1067,7 @@ Public Class clsPlcIf
     ''' </summary>
     ''' <param name="TagName"></param>
     ''' <param name="Value"></param>
-    Public Sub AnalogPlcWrite(TagName As String, ByVal Value As Short)
+    Public Sub AnalogPlcWrite(TagName As String, ByVal Value As Single)
         With AnalogTag.TagData(TagName)
             Dim PlcAdress As String = .Address
             'アナログデータをPLC書込データに変換
